Merged PR 420231: Making ConfidenceBoundUtilityFunction more intuitive
Andreas pointed out that to the user of the BayesianOptimizer the interaction between these parameters is confusing and can lead to subtle bugs:
 {maximize, minimize} x {upper_confidence_bound, lower_confidence_bound}.

As one would often use upper_confidence_bound for maximization and lower_confidence_bound for minimization .

However, these parameters are set in different places (OptimizationProblem vs. OptimizerConfig) and entirely non-intuitive. So this PR renames them to upper_confidence_bound_on_improvement and lower_confidence_bound_on_improvement and changes the math accordingly. This way, when the user switches from maximizing, to minimizing, she/he/they doesn't/don't have to change the optimizer config. This rename is in-line with the names of other utility functions: Probability of Improvement, Expected Improvement.