Merged PR 421281: Switching from sample variance to mean variance
Using sample variance instead of mean variance to compute the confidence intervals makes them not shrink as we add more observations.

While this measure can still be useful down the road to detect noisy-TV type of problems, using mean variance makes more sense for the problems we are solving now.