Merged PR 421755: Ensure default parameters are copied
This PR ensures default parameters are copied and not accidentally modified by the user.

I'm also introducing a long-haul test for BO, which was the original motivation. Happy to split that off into a different PR.

I wasn't sure where to put the DefaultConfigMeta helper, lmk if there's a better place.

There's two reasons this is tricky: DEFAULT is used on the class, but properties are instance-level, so we need a meta-class. Also, because of the way the dictionaries are instantiated, we can't make copies of nested points easily, so we need to use deepcopy.

This solution still has the issue that we're using mutable objects as default parameters. If a user would modify ``_DEFAULT`` for some reason, the default instantiations would change.