Merged PR 427208: Corrected Numerical Instability In Variance Computation
This run revealed an issue in variance computation:

https://msdata.visualstudio.com/Database%20Systems/_build/results?buildId=21291699&view=results

Specifically, a variance value of -7.450581e-09 made the sqrt throw. This PR corrects it by adding a small constant to the variance in the Random Forest.

<s>At the moment this PR also includes a beefed up test that revealed the problem. Once that test passes we will have good evidence that the problem is resolved. I will restore the test to its original state and check in.</s>

The test ran all night redoing the computation on every iteration. It did not throw so I conjecture that the bug is fixed.