Merged PR 427161: Glow Worm Utility Function Optimizer
The main portion of this PR is the implementation of the Glow Worm Swarm Optimization algorithm as an alternative to the RandomSearchOptimizer.

You will find its implementation in: ```\mlos\Optimizers\ExperimentDesigner\UtilityFunctionOptimizers\GlowWormSwarmOptimizer.py```

The first part of this post has a good description of the algorithm: https://www.hindawi.com/journals/mpe/2016/5481602/

Note that our implementation is aware of hierarchical hypergrids and essentially will split up the glow worms so that each group operates on a single, consistent space (i.e. glow worms will not in general move from LRU to MRU implementation except for random restarts).

There is still plenty of low-hanging fruit (like watermelon size) in the implementation. We should get a big improvement by remembering the best parameters from previous calls to GlowWormSwarmOptimizer.suggest() and reusing them as the starting positions for subsequent calls to suggest (with some randomness of course). Other possible improvements include: varying the step size, varying the number of iterations, implementing better stopping criteria. But this PR is already big, so we can do that later.

In order to plug in this optimizer I needed to standardize the UtilityFunctionOptimizer interface, hence the base classes: UtilityFunctionOptimizer and UtilityFunction were born. I also needed an adapter to translate points into a unit continuous search space. As I was going along, I also fixed a few small bugs.

Lastly, the new adapters and the existence of the glow worm optimizer make it pretty trivial to implement a random-near-incumbent strategy that Ed says is hard to beat. That would be really something - having 3 utility function optimizers, 2 surrogate models, 2 utility functions would allow us to instantiate 12 distinctly different optimizers and this number will grow super-linearly with every new implementation.