Merged PR 429225: Renamed translate to project in HypergridAdapters
Since the HypergridAdapters actually project points from one space to another, I thought a name change would be good. Especially that translation means "moving a point by some vector".

ALSO: I added a little default behavior of being able to automatically project points if the authors of a HypergridAdapter derived class decides it's not worth their while to implement it by hand.

The usage of HypergridAdapters as a wonderful data pre-processing tool has not caught up yet, so this change affects only my code and I have tested it :) But I encourage you to start using the adapters, because they are a reusable, clean and effective way of translating configurations from what the user wants to see, to what the models can compute :)

Here are some ideas for useful features:
* a PCA adapter - to preprocess all of those context values we are about to start pumping into our models
* a one-hot-encoding adapter - as a reusable way to deal with categorical dimensions
* a bootstrapping adapter
* a column-dropping adapter
* a logit adapter - to preprocess the objectives that have finite ranges