Fixup documentation links (#61)
A bunch of work for fixing up some broken documentation links and improving the pipelines to try and prevent that in the future.

Uses a fake anchor hack discussed previously to identify which links, when rendered in github pages, should be sent back to the main github repo page, and which should remain relative, though suitably fixed up.

Adds a Makefile for building the site both locally (for testing) as well as in the pipelines.

This also fixes a an issue with the python_api code being copied to the wrong location so it wasn't being updated.

Adds a workflow to test for broken links in the plain markdown files as well.

None of them check for faulty anchors.

Partial work on providing the ability to execute the notebooks prior to markdown rendering as well (there's a bug that seems to be related to the recent shell=True fixup from #46 I haven't addressed there yet though).

For that we also needed the ability to install all of the necessary pip modules. I tried to add a script to avoid use of requirements.txt and instead use the conda environment.yml without actually making use of conda (kind of a hack for the moment).

Eventually I intend to replace that with proper docker container build image reuse.