Merged PR 431690: Started Working on Optimizer Evaluation Toolkit
This is the first of the PRs introducing the optimizer evaluation toolkit.  Much of the toolkit's description is in the OptimizerEvaluationTools notebook so please refer to it for details.

Broadly speaking, I put together a notebook that demonstrates how the optimizer evaluation tools can be used. So far we only keep track of goodness of fit, but I plan to add optima, and utility function optimziers to the toolkit too. I mostly try to replicate what Ed and Andreas do to understand the optimizers' performance and automate it as much as possible. I even used a bunch of Ed's visualizations :)

**About the framework**: most of the synthetic functions we have used so far, now derive from SyntheticFunctionBase - basically exposing an interface to evaluate_point(), evaluate_dataframe(), and expose: parameter_space, output_space, and context. These functions are produced by the SyntheticFunctionFactory. Ideally, we will build up a portfolio of interesting functions, and @<Andreas Mueller> said he could point us to already existing libraries.

Additionally, this PR introduces a ComponentConfigStore class, which we discussed in the Mlos channel. It is meant as a replacement for all of the XYZConfig classes. So far I have only instantiated two stores: ObjectiveFunctionConfigStore, and DecisionTreeConfigStore and once we are happy with the interface, and functionality, I'll replace all occurences of XYZConfig.