initial work to binplace binaries on linux
This provides "make install" style targets to binplace binaries to target/bin/{Release,Debug}/ for easier execution.

Note that SettingsRegistry assemblies were already being placed there.

With the current Mlos.Agent lookup behavior, this means that we can now
run

```sh
cd target/bin/Release
./Mlos.Agent.Server ./Mlos.UnitTest
```

Instead of having to specify
1. the path to the dotnet tool
2. the paths to the Mlos.Agent.Server assembly
3. the path to the target executable
4. the MLOS_SETTINGS_REGISTRY_PATH environment variable

To do this, we provide an Mlos.NetCore.Publishing.targets file that
provides a new target "PublishProjectBinary" that in turn runs an
appropriate "dotnet publish" command for the given platform.

This method requires the fewest changes to the existing build system
since we can specify the RuntimeIdentifier in the command line instead
of in all of the build files.

To allow recursive operation for projects that don't include that
target, we let that target be a no-op most of the time.

Next, we hook that into the existing Makefile and cmake wrappers as
well.

TODO:

- tests
- documentation