2020/september/optimizing predict performance (#92)
This change improves runtime of the DecisionTreeRegressionModel.predict() function.

The main idea is to use column-wise operators so that pandas and numpy can vectorize them. Tracing reveals 2x-200x improvement in filtering out invalid rows depending on the complexity of the hypergrid and the number of rows in the dataframe.