Extending Optimum Functionality (#89)
Added a class enumerating all supported definitions of "optimum" and implemented three of them:
* PREDICTED_VALUE_FOR_OBSERVED_CONFIG
* LOWER_CONFIDENCE_BOUND_FOR_OBSERVED_CONFIG
* UPPER_CONFIDENCE_BOUND_FOR_OBSERVED_CONFIG
Users can select which of the definitions to use based on their risk profiles. For the confidence bound based optima, the users can specify the probability of type 1 error (alpha) to indicate their required level of certainty.