Merged PR 441469: The target process creates the config shared memory.
Only the target process creates the config shared memory.

Global shared memory now contains the information about registered memory regions and registered settings assembly.

Now the target process creates the config shared memory, where it create it with the desired size and name.
(Configuration is not exposed yet, but changing it does not require changes in the agent and the target process).

We do support only one shared config memory region, but now this can be extended.

Fix a bug when shared maps are not cleanup in case of initialization failure.
Implementation of dictionary lookups is moved to SharedConfigDictionaryLookup struct.

Next PR: Remove static fields from MlosContext.