quick tweak to support updating dotnet on a runner
Works around some CI issues with "dotnet tool" commands for Cake build with
older dotnet versions on older Ubuntu 16.04 ADO runners and similar issues for
cmake.

TODO: This will need additional fixes in the future.

We still have the setup-cmake.sh, setup-dotnet.sh type of scripts which can
install local copies, but we've started the process of stripping use of those
out of the build system.  One benefit of those is that they also allow us to
fetch (pin) an explicit version (though that was probably a security issue) and
use that version even if the default system repo lists don't have a newer
version for us.

A more complete fix will be to include the logic to check for an appropriate
version in the existing system repos (not just existence of the package), and if
it doesn't exist add the necessary package repos as well.