Slight tweaks for non-pipeline builds on the ADO repo
This works around the fact that GitVersionTask doesn't support pre-release tags
(the only kinds we have for the "master" branch in ADO atm), when attempting to
compute the current "version" as an offset from the last tag in Mainline mode.

Instead, we detect when we're on the ADO branch by looking for the ADO
`.pipelines\pipeline.windows.yml` file and disable GitVersionTask in that case.

Also, since Mlos.NetCore.Components.Packages.proj is not a .csproj, it doesn't
include the `build/version/Mlos.NetCore.Version.props` file, so we move those
property definitions to `build/Mlos.Common.props`.

Note that this doesn't prevent that project from packaging things when built on
ADO (though it isn't included by default in `dirs.proj`), so one could still
test the `ExternalIntegrationExample.SettingsRegistry.csproj` manually, however
the versions of the `.nupkg` files it produces will currently always be `1.0.0`
so some additional effort may be required to replace previously restored
packages using that version.