Create Nuget Packages for C# (#177)
This is the first step in a line of work aimed towards making MLOS more natively consumable by external projects.

The basic idea here is that we use the standard .NET.Sdk method of packaging each of the relevant `.csproj`s as their own `.nupkg`.
This let's us rely on `dotnet` to automatically track `PackageReference` and `ProjectReference` dependencies and generate them into the `.nuspec` that produces the `.nupkg`.

Since `.nupkg` files are considered immutable, we have to create unique version numbers for each of them.
To do this, I added [`GitVersionTask`](https://github.com/GitTools/GitVersion), but only for the OSS builds.

Typically, GitVersion expects one to use GitFlow or GithubFlow workflows with develop and release branches.
We are currently only using a Mainline workflow model, so the `GitVersion.yml` file is configure as such.
The versions produced use SemVer by default, so when building on a branch, we get a version like the following:
`Mlos.SettingsSystem.CodeGen.0.1.1-codegen-nuget-for-external-project-integration.33.nupkg`
Where 
- the `0.1.1` portion is taken as the number of commits on the mainline branch since the last v0.1.0 tag found in the repo
- the `-codegen-nuget-for-external-project-integration` is the name of the branch
- and the `.33` is the number of commits since that branch point.
In effect, we can use `git tag` to version our `x.y.z` releases, at least until we start adding `release/0.1.x` style servicing branches.

For the moment, the `.nupkg` files are not signed and not published (e.g. to nuget.org).  Instead, we produce them into the `target/pkg/$(Configuration)/` tree whenever either the new `Mlos.NetCore.Components.Packages.proj` is build or else when `dotnet pack` is run on some `.csproj`.

The `ExternalIntegrationExample.SettingsRegistry.csproj` (which otherwise uses no Mlos build rules and acts as a standard `.NET.Sdk` example) is used for testing and is configured to pull from that tree (for local testing) as well.

The next stage of work (separate PR) is to configure a set of `cmake` rules using `FetchContent` to grab the C++ components, build these packages and reference them when building the rest of a C++ project and its dependencies.

TODO:
- [x] Enable the test project and package building in CI
  - [x] for Windows
  - [x] for Linux
- [x] Test VisualStudio build
- [x] Verify CDPX/CoreXT builds (!473121)

WONTFIX:
- Leaving actual publishing/signing to nuget.org for a future PR