Computing Pareto (#181)
This PR introduces a simple algorithm for computing a pareto frontier, given an objectives dataframe and a corresponding OptimizationProblem.

You will find the algorithm in ParetoFrontier. This class will later be extended to maintain a cache of the latest pareto, and to iteratively add to it as more observations roll in. It will be owned by a BayesianOptimizer.