External cmake project integration (#193)
Work to make it easier/possible to integrate the C++ components of MLOS with an external C++ project (e.g. LevelDB, MySql, etc.) via `cmake`.

- Fixes some issues with our own `cmake` config assumptions.
- Adds examples patterns and documentation for using in an "external project" (currently just a minimal example).
- Hooks those into the CI pipelines.

This branch was split off of the work from #177 (since merged) to build NuGet packages for some of the C# components, so many of the original commits are shared.