Integrating multi objective random forest into the optimizer framework (#212)
This is the next batch of code getting us towards multi-objective optimization.

PAST: Previously we have been able to register multiple objectives with the optimizer, but the optimizer wouldn't use them at all.
PRESENT: With this PR the optimizer will instantiate a regression model for each of the objectives, but it doesn't yet use the predictions for anything.
FUTURE: The next steps will be to write a few multi-objective utility functions so that the optimizer can use them to suggest new configurations.

The chief goal of this PR is to integrate the MultiObjectiveHomogeneousRandomForest into the framework.

Note that there is no change to the Optimizer's interface. The prediction of the first objective is returned, and all the others are thrown out.