Random Near Incumbent Optimizer (#226)
This PR introduces a third utility function optimizer: RandomNearIncumbent. Its implementation is described in detail in RandomNearIncumbentOptimizer.py.

The interface between the UtilityFunctionOptimizer subclasses and the Experiment designer has been changed. Now, if a given UtilityFunctionOptimizer is unable to produce a guided suggestion, it raises an Exception and the ExperimentDesigner can decide how to handle it and perhaps explicitly generate a random suggestion. All existing utility function optimizers have been updated and TestUtilityFunctionOptimizers.py has been extended to cover this new case.

Testing: I ran pyest with code coverage and got 100% coverage for both RandomSearchOptimizer and GlwoWormSwarmOptimizer. For RandomNearIncumbentOptimizer the score is 93%, but the only lines missing are really defensive checks, that raise an exception if something is wrong at the end of the suggest() method.

I also temporarily, made the RandomNearIncumbent the default optimizer and ran all unit tests with it locally. It is considerably slower than RandomSearchOptimizer so I reverted the to the old default having verified that it works.