Moving Optimizer Store out of Optimizer Microservice (#240)
The goal of the upcoming PRs is to split gRPC into two Microservices:

1. OptimizerMicroservice - implementing only very few interfaces so that we can have optimizer clients in multiple languages.
2. OptimizerMonitoringMicroservice - implementing all of the other APIs currently defined in the .proto file.

gRPC forces us to make two separate Service classes but their objects will need to be able to reference the same optimizers. So step 1 is to move the OptimizerStore to its own class so that it could be referenced by both services. Down the road, we should evolve the OptimizerStore to use a Database or filesystem or MLFlow. For now it stays in memory.