Defining all Hypergrid types in gRPC (#246)
The goal of this PR is to define all Hypergrid types in proto3 and build an encoder/decoder pair to support it.

Those messages are not yet used by the gRPC service, as this would break the existing clients. A subsequent PR will implement encoders/decoders in C# at which point we will be able to switch the gRPC services to the new message types.