Merged PR 573: Document quick start guide for mlos_bench
This PR updates the README.md of `mlos_bench` with the steps needed to run a user's own JSON configuration with `main.py`.

Additionally, a small fix is included in the template `config-example.json` for specifying the corresponding `vmName` when establishing the VM, otherwise we end up always using the default  value of `osat-linux-vm` provided in the ARM template.

Related work items: #485