Merged PR 576: Azure VM controls
This PR implements more VM control methods in `azure_services.py` like:
- Stop
- Deprovision
- Reboot

These methods are also updated to be used in the teardowns for `azure_os.py` and `azure_vm.py`.
`main.py` is updated to now teardown the composite environment at the end of the experiment, and we can observe the VM shutting down and being de-provisioned in Azure Portal.

Related work items: #469