Merged PR 862: Add devcontainer support and other dev environment improvements
This PR is about general dev environment improvements.

Adds devcontainer support to help automatically prepare a dev environment inside either vscode of github codespaces.

The container is build, tested, and published to an ACR in the MlosCore RG during the pipeline for easy reuse and quick dev env boot strapping and reproducible environments both locally and remotely.

It also
- Enables parallel Makefile builds
- Adds additional helpful extensions to the vscode (e.g. copilot)
- Moves some pip package installs to conda for better reproducibility
- Fixes some issues regarding PYTHONPATH adjustment requirements for pytest and pylint.
- And, since I noticed a config for it in vscode, I added pydocstyle checks, though they aren't configured for blocking atm since there's lots of complaints.  Can decide whether to fix/force those later.