Devcontainer port conflict tweaks (#296)
Windows (or other platforms) may listen on port 80 and cause conflicts when WSL2 tries to forward ports for both sides so that `localhost` resolution behaves as expected despite the VM technically being a separate host.
When this happens the devcontainer can fail to start due to in ability for nginx docs container to listen on port 80.
To workaround this we simply tell it to listen on port 81 instead.  The devcontainer forwards the port to a different one on the host anyways.

Additionally, we document some commands for cleaning up stale container images in the ACR service to avoid excessive storage usage.