Fixes for TunableGroups merging logic (#327)
Splitting more work off of #313.

In that PR we'd like to be able to make each Environment independently loadable.
To do this we need to be able to allow them to reference their Tunables.
However, other related Environments (e.g. local vs. remote), may also need to reference those same Tunables.
Moreover, we want both to obtain the same value.
To do this we would like the parent TunableGroups for the CompositeEnvironment for an Experiment to handout *references* rather than *copies* of Tunables to the child Environments.

This work adds that logic and some basic tests for it.

In essence, we make passing TunableGroups to an Environment expected, even if the first copy may be an empty collection.  We consider this one the "root" or "parent" copy.
After that, any additional Tunables loaded by subsequent Environments get `merge`d (not `update`d, so no overwriting) into the "parent" and the child gets a "view" of a `subgroup` of those.
In this way, when the Optimizer assigns new values to a Tunable, it is reflected at all levels.

Caveats:
- Currently, the `merge` logic throws an error if any overlapping CovariantTunableGroups are non-equal in any way (e.g. extra param in one vs. another).

Leaving these for future work:
- [ ] Add tests for nested CompositeEnvironments to ensure that we pass the "parent" to multiple levels deep, not just the first one.