SMAC Fixups (#431)
When "start_with_defaults" is in effect, we `register` without doing a `suggest`.
Internally, the SmacOptimizer keeps a map of prior TrialInfo received from `suggest` for registering configs.
Since this didn't exist, we create a blank one. However, the blank one was lacking the `seed` info, which SMAC expects.

This PR adds that info and enables the test for it.