Add unit tests for the tunable groups selection (#440)
Also, fix the confusing behavior of the environment: we did not assign
the new values to the environment's tunables after `.setup()`. That is
ok most of the time as we never use the tunables alone when setting up
the environment (we merge them into `self._params` along with
`const_args` and globals), but it is still a bit confusing to not update
the tunables on setup.