Make "include_tunables" work for composite environments (#453)
There is a bug in `CompositeEnv` constructor that prevents the parent's
tunables to be pushed down into the child environments. In particular,
the tunables from `"include_tunables"` in the composite environments are
not available to all children without mixing them into the
`CompositeEnv._tunable_params`.

That issue makes impossible some perfectly legitimate configs, e.g.,
```jsonc
{
    "name": "Setup Ubuntu VM",
    "class": "mlos_bench.environments.composite_env.CompositeEnv",
    "include_tunables": ["environments/os/linux/runtime/linux-runtime-tunables.jsonc"],
    "config": {
        "children": [
            {
                "name": "Generate Linux kernel parameters for Ubuntu",
                "class": "mlos_bench.environments.LocalFileShareEnv",
                "config": {
                    "tunable_params": ["$linux-runtime"],
                    // ... etc.
```
Note that `"include_tunables"` is in the composite env whereas
`"tunable_params"` is one level below it.

We need that kind of configurations for local-remote-local composites
where all three child environments share the same set of tunables. This
PR fixes the logic for propagating the tunables' data up and down the
environment hierarchy