bugfix: do not expand the script path in run/setup/teardown if it is not executable (#459)
Checking for an actual executable is a bit tricky on Windows, so we'll
do it properly later; (also, it is ok to have X bits not set for Python
scripts). For now, add some minimal checks to prevent `.` (as in `.
script.sh`) from expanding into a current directory path instead of
being treated as `source script.sh` in bash.
Also, remove extra newlines when logging the script's stdout/stderr.