Expand dollar variables in the globals (#490)
This feature greatly simplifies our configs for complex multi-VM setups.

P.S. I cannot come up with any simple and elegant unit tests for this
feature. Technically, we can define some dollar variables in
`global_test_config.jsonc` and check their values in the log in
`launcher_test.py`, but we don't have that many parameters in `MockEnv`
and `MockOptimizer` that we can expand. I'll think about it and do it in
the next PR.