Fix the nested service inclusion issue + add unit tests (#551)
**UPDATE:** Fixed! The problem was with the `.register()` method
invocation in the base class constructor, so the service can inherit
methods from the parent service. That call overrode methods of the
service object being created with references to the parent's methods. As
a result, we were losing references to the local methods before we could
register them. This PR is a workaround for the issue. We will think of a
more elegant way of doing it later.

-------------------

Debugging and fixing the issue described in #530

**TODO:**
* [x] Fix the original issue from #530
* [x] Add unit tests to verify the problem
* [x] ~Fix the JSON schema to accept the `CompositeEnv` fixture from
unit tests~
* [x] Resolve `mypy` issues with accessing `.temp_dir_context()`