Fixup issues with VSCode and pytest working directory paths in different environments
Previously #553 solved this for pytest inside VSCode for WSL, devcontainer within WSL, and Windows directly dev environments by looking at the `$VSCODE_CWD` environment variable that the Pytest VSCode extension sets to note the difference between the expected environment variable `$PWD` and the result of `os.getcwd()`.

However, when the devcontainer is launched from Windows, the VSCode pytest plugin takes a different approach and launches a proxy server in WSL to run the tests.
In that mode, `$VSCODE_CWD` is still set but is no longer the path used.

To avoid fighting with all these different scenarios, we change the test to set `$CUSTOM_PATH_FROM_ENV` via `os.environ` prior to the start of the test, and then make sure that it's expanded when we use it inside `global_test_config.jsonc`

Closed #575