Address some possible issues with asyncio usage.
When the the EventLoopContext and SshService context exits and cleanup code starts, it may be possible that there are still outstanding tasks in the event loop, in which case it can take a moment for the tasks to finish.

In particular, if one of those tasks is a coroutine that holds a lock, possibly that lock won't be released until the coroutine finishes.

If a new context is then started, and a new copy of that coroutine attempts to acquire the shared lock, it will throw an error about the lock already being held by a different loop.

To address this, we make a handful of changes:

- warn if the lock is still held when context exit code happens for the SshClientCache (mostly for future debugging help)
- in the EventLoopContext exit code, we add a timeout to wait for the loop to finish running outstanding tasks
- in the SshService, we call the event loop cleanup code prior to the SshClientCache (so that any outstanding requests there can be sure to release any locks held)