CI tweak: use "release" event instead of "tag push" for pypi publishing (#631)
Minor tweak on #626: instead of acting on tag push events, we only act
on "release" events. This gives us a chance to fix bad tag pushes (which
otherwise don't have much gated checking on them) prior to attempting to
publish a version on pypi, which cannot be overwritten (only
superseded).