Allow loading configs from JSON strings and test with new docstrings for tunables (#890)
# Pull Request

## Title

Allow loading configs from JSON strings not just files and add docs for
Tunables to demonstrate and test it.

---

## Description

Adds basic support for loading a JSON config string directly.

The main purpose is to be able to show examples in docstrings that are
testable (so that they don't become stale examples).

- [x] Add basic example of usage and testing in a tunables docstring
example.
- [x] Add `pytest` unit tests for config string loading as well (we
already have file loading tests).
- Closes #881 

---

## Type of Change

- ✨ New feature
- 📝 Documentation update
- 🧪 Tests

---

## Testing

- [x] Adds additional unit tests.

---

## Additional Notes (optional)

Split out from #887 

---