Introduce TrialRunner Abstraction (#720)
This is another step in adding support for parallel trial execution
#380.

Here we separate out the running of an individual trial to a single
class - TrialRunner.

Multiple TrialRunners are instantiated at CLI invocation with the
`--num-trial-runners` argument. 
Each TrialRunner associated with a single copy of the root Environment
and its Services (important for eventual parallelization reasons), and
made unique by means of a unique `trial_runner_id` value that's also
included in that Environment's global_config.

In future PRs we will add:
- New Scheduler implementations to run TrialRunners in parallel.
- Async polling of status results in each TrialRunner independently.

---------

Co-authored-by: Sergiy Matusevych <sergiym@microsoft.com>
Co-authored-by: Sergiy Matusevych <sergiy.matusevych@gmail.com>